+++ internal/model/model.go
@@ -1,0 +1,8 @@
+/*
+Package model provides describes the internal types and their behaviors
+*/
+package model
+
+// init function intiailizes the model package
+func init() {
+}
