+++ internal/model/reminder_data.go
@@ -1,4 +1,4 @@
-package models
+package model
 
 import (
 	"encoding/json"
