--- internal/models/models.go
@@ -1,8 +1,0 @@
-/*
-Package models provides describes the internal types and their behaviors
-*/
-package models
-
-// init function intiailizes the models package
-func init() {
-}
