+++ scripts/go_test
@@ -14,5 +14,5 @@ run_go_test() {
 
 # run tests
 run_go_test "pkg/utils/"
-run_go_test "internal/models/"
+run_go_test "internal/model/"
 run_go_test "cmd/reminder/"
